fix(tests/irsbot): moving tests from graphql.expenses to irs.bot
