--- server/constants/comments.js
@@ -1,8 +1,0 @@
-
-export const MAXIMUM_AMOUNT_TO_SEND_FORM_COMMENT = 60000; // U$ 600.00 total amount allowed without form as of July 2018
-export const expensesFormData = {
-  FORM_EXPENSE_COMMENT_HTML: '<p>You have now been paid $600 or more through Open Collective, which' +
-    ' means we need to ask you fill out a tax form. For more info,' +
-    ' <a href="https://github.com/opencollective/opencollective/wiki/Submitting-Expenses#taxes">' +
-    'click here see the help wiki.</a></p>'
-};
